fix: redirectio to bad_url_path
moved get_bad_url_redirect_path from core.py to _app.py
